fix: Activate user
