+++ src/features/users/PageUsers.tsx
@@ -16,31 +16,15 @@ import {
   IconButton,
   LinkBox,
   LinkOverlay,
-  Menu,
-  MenuButton,
-  MenuDivider,
-  MenuItem,
-  MenuList,
-  MenuProps,
-  Portal,
   Text,
   Wrap,
   WrapItem,
 } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import {
-  FiCheckCircle,
-  FiEdit,
-  FiPlus,
-  FiRefreshCw,
-  FiTrash2,
-  FiXCircle,
-} from 'react-icons/fi';
+import { FiPlus, FiRefreshCw } from 'react-icons/fi';
 import { Link } from 'react-router-dom';
 import { useSearchParams } from 'react-router-dom';
 
-import { ActionsButton } from '@/components/ActionsButton';
-import { ConfirmMenuItem } from '@/components/ConfirmMenuItem';
 import {
   DataList,
   DataListCell,
@@ -49,7 +33,6 @@ import {
   DataListRow,
 } from '@/components/DataList';
 import { DateAgo } from '@/components/DateAgo';
-import { Icon } from '@/components/Icons';
 import { Page, PageContent } from '@/components/Page';
 import {
   Pagination,
@@ -59,130 +42,11 @@ import {
   PaginationButtonPrevPage,
   PaginationInfo,
 } from '@/components/Pagination';
-import { useToastError, useToastSuccess } from '@/components/Toast';
 import { AdminNav } from '@/features/admin/AdminNav';
 import { UserStatus } from '@/features/users/UserStatus';
-import {
-  useUserList,
-  useUserRemove,
-  useUserUpdate,
-} from '@/features/users/service';
-import { User } from '@/features/users/types';
+import { useUserList } from '@/features/users/service';
 
-type UserActionProps = Omit<MenuProps, 'children'> & {
-  user: User;
-};
-
-const UserActions = ({ user, ...rest }: UserActionProps) => {
-  const { t } = useTranslation(['common', 'users']);
-  const toastSuccess = useToastSuccess();
-  const toastError = useToastError();
-  const userUpdate = useUserUpdate({
-    onSuccess: ({ activated, login }) => {
-      if (activated) {
-        toastSuccess({
-          title: t('users:feedbacks.activateUserSuccess.title'),
-          description: t('users:feedbacks.activateUserSuccess.description', {
-            login,
-          }),
-        });
-      } else {
-        toastSuccess({
-          title: t('users:feedbacks.deactivateUserSuccess.title'),
-          description: t('users:feedbacks.deactivateUserSuccess.description', {
-            login,
-          }),
-        });
-      }
-    },
-    onError: (_, { activated, login }) => {
-      if (activated) {
-        toastError({
-          title: t('users:feedbacks.activateUserError.title'),
-          description: t('users:feedbacks.activateUserError.description', {
-            login,
-          }),
-        });
-      } else {
-        toastError({
-          title: t('users:feedbacks.deactivateUserError.title'),
-          description: t('users:feedbacks.deactivateUserError.description', {
-            login,
-          }),
-        });
-      }
-    },
-  });
-
-  const activateUser = () => userUpdate.mutate({ ...user, activated: true });
-  const deactivateUser = () => userUpdate.mutate({ ...user, activated: false });
-  const isActionsLoading = userUpdate.isLoading;
-
-  const userRemove = useUserRemove({
-    onSuccess: (_, { login }) => {
-      toastSuccess({
-        title: t('users:feedbacks.deleteUserSuccess.title'),
-        description: t('users:feedbacks.deleteUserSuccess.description', {
-          login,
-        }),
-      });
-    },
-    onError: (_, { login }) => {
-      toastError({
-        title: t('users:feedbacks.deleteUserError.title'),
-        description: t('users:feedbacks.deleteUserError.description', {
-          login,
-        }),
-      });
-    },
-  });
-  const removeUser = () => userRemove.mutate(user);
-  const isRemovalLoading = userRemove.isLoading;
-
-  return (
-    <Menu isLazy placement="left-start" {...rest}>
-      <MenuButton
-        as={ActionsButton}
-        isLoading={isActionsLoading || isRemovalLoading}
-      />
-      <Portal>
-        <MenuList>
-          <MenuItem
-            as={Link}
-            to={`/admin/users/${user.login}`}
-            icon={<Icon icon={FiEdit} fontSize="lg" color="gray.400" />}
-          >
-            {t('common:actions.edit')}
-          </MenuItem>
-          {user.activated ? (
-            <MenuItem
-              onClick={deactivateUser}
-              icon={<Icon icon={FiXCircle} fontSize="lg" color="gray.400" />}
-            >
-              {t('common:actions.deactivate')}
-            </MenuItem>
-          ) : (
-            <MenuItem
-              onClick={activateUser}
-              icon={
-                <Icon icon={FiCheckCircle} fontSize="lg" color="gray.400" />
-              }
-            >
-              {t('common:actions.activate')}
-            </MenuItem>
-          )}
-          <MenuDivider />
-          <ConfirmMenuItem
-            icon={<Icon icon={FiTrash2} fontSize="lg" color="gray.400" />}
-            onClick={removeUser}
-          >
-            {t('common:actions.delete')}
-          </ConfirmMenuItem>
-        </MenuList>
-      </Portal>
-    </Menu>
-  );
-};
+import { UserActions } from './UserActions';
 
 export default function PageUsers() {
   const { t } = useTranslation(['users']);
